Fixed and moved test
